product_review_system/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.0
django-cors-headers>=4.3.0
django-debug-toolbar>=4.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyJWT>=2.8.0
pytz>=2023.3
sqlparse>=0.4.4
uritemplate>=4.1.1
inflection>=0.5.1
packaging>=21.0
